Creacion De variable email

+++ infrastructure/driven-adapters/jpa-repository/src/main/java/co/com/cidenet/employees/jpa/JPARepositoryAdapter.java
@@ -4,11 +4,14 @@ import co.com.cidenet.employees.jpa.data.EmployeeData;
 import co.com.cidenet.employees.jpa.helper.AdapterOperations;
 import co.com.cidenet.employees.model.employee.Employee;
 import co.com.cidenet.employees.model.employee.gateways.EmployeeRepository;
+import co.com.cidenet.employees.model.exception.BusinessException;
 import org.reactivecommons.utils.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @Repository
 public class JPARepositoryAdapter extends AdapterOperations<Employee, EmployeeData, Long, JPARepository>
@@ -30,12 +33,26 @@ public class JPARepositoryAdapter extends AdapterOperations<Employee, EmployeeDa
     }
 
     @Override
-    public void deleteEmployee(Long id) {
-         repository.deleteById(id);
+    public void deleteEmployee(Long id) throws BusinessException {
+        Employee employee = super.findById(id);
+        if(Objects.isNull(employee)){
+           throw new BusinessException(404,"No se encontro el empleado " + id);
+        }
+
+        repository.deleteById(id);
     }
 
     @Override
     public List<Employee> getAll() {
         return super.findAll();
+    }
+
+    @Override
+    public List<Employee> findByNameAndAndLastNameAndCountry(String name,String lastName, String country) {
+        List<Employee> employees = new ArrayList<>();
+        repository.findByNameAndAndLastNameAndCountry(name,lastName,country)
+                .forEach(x -> employees.add(mapper.map(x,Employee.class)));
+
+        return employees;
     }
 }
